fix<bug>: fix automatic scrolling error
The animateTo method was throwing an exception because on the creation of the first text field group the method was running before the group is rendered. So the fix was to wrap that method with a addPostFrameCallback method. This ensured that animateTo is only being called after render is complete